New migrations + sql for settings + fixed some bugs

+++ database/migrations/2019_05_31_155500_create_achievement_types_table.php
@@ -19,7 +19,7 @@ class CreateAchievementTypesTable extends Migration
             $table->string('type');
             $table->string('stage');
             $table->string('result');
-            $table->integer('score');
+            $table->float('score');
             $table->timestamps();
         });
     }
